+++ src/snowflake/snowpark/dataframe_reader.py
@@ -83,6 +83,7 @@ from snowflake.snowpark.dataframe import DataFrame
 from snowflake.snowpark.exceptions import (
     SnowparkSessionException,
     SnowparkDataframeReaderException,
+    _SnowparkDataSourceNonRetryableException,
 )
 from snowflake.snowpark.functions import sql_expr, col, concat, lit, to_file
 from snowflake.snowpark.mock._connection import MockServerConnection
@@ -2003,7 +2004,13 @@ class DataFrameReader:
                             f"Cancelled a remaining data fetching future {future} due to error in another thread."
                         )
 
-            if isinstance(exc, SnowparkDataframeReaderException):
+            if isinstance(
+                exc,
+                (
+                    SnowparkDataframeReaderException,
+                    _SnowparkDataSourceNonRetryableException,
+                ),
+            ):
                 raise exc
 
             raise SnowparkDataframeReaderException(