+++ tests/integ/datasource/test_mysql.py
@@ -14,6 +14,7 @@ from snowflake.snowpark._internal.data_source.drivers.pymsql_driver import (
     PymysqlTypeCode,
 )
 from snowflake.snowpark._internal.data_source.utils import DBMS_TYPE
+from snowflake.snowpark.exceptions import _SnowparkDataSourceNonRetryableException
 from snowflake.snowpark.types import StructType, StructField, StringType
 from snowflake.snowpark.exceptions import (
     SnowparkDataframeReaderException,
@@ -303,6 +304,18 @@ def test_unsupported_type():
         [("test_col", "unsupported_type", None, None, 0, 0, True)]
     )
     assert schema == StructType([StructField("TEST_COL", StringType(), nullable=True)])
+
+
+def test_mysql_non_retryable_error(session):
+    with pytest.raises(
+        _SnowparkDataSourceNonRetryableException,
+        match="You have an error in your SQL syntax",
+    ):
+        session.read.dbapi(
+            create_connection_mysql,
+            table=TEST_TABLE_NAME,
+            predicates=["invalid syntax"],
+        )
 
 
 def test_session_init(session):